+++ projects/_01_ingest/cqc_pir/jobs/validate_pir_cleaned_data.py
@@ -6,9 +6,7 @@ os.environ["SPARK_VERSION"] = "3.3"
 from pyspark.sql.dataframe import DataFrame
 
 from utils import utils
-from utils.column_names.ind_cqc_pipeline_columns import (
-    PartitionKeys as Keys,
-)
+from utils.column_names.ind_cqc_pipeline_columns import PartitionKeys as Keys
 from utils.validation.validation_rules.pir_cleaned_validation_rules import (
     PIRCleanedValidationRules as Rules,
 )
