+++ projects/_01_ingest/unittest_data/ingest_test_file_data.py
@@ -652,3 +652,23 @@ class CleanCQCPIRData:
         ("1-1199876096", "N", date(2023, 6, 1), date(2023, 5, 24)),
         ("1-1199876096", "Y", date(2023, 6, 1), date(2023, 5, 24)),
     ]
+
+
+@dataclass
+class ValidatePIRRawData:
+    raw_cqc_pir_rows = [
+        ("1-000000001", "20240101", 10),
+        ("1-000000002", "20240101", 10),
+        ("1-000000001", "20240109", 10),
+        ("1-000000002", "20240109", 10),
+    ]
+
+
+@dataclass
+class ValidatePIRCleanedData:
+    cleaned_cqc_pir_rows = [
+        ("1-000000001", date(2024, 1, 1), 10, "Y"),
+        ("1-000000002", date(2024, 1, 1), 10, "Y"),
+        ("1-000000001", date(2024, 1, 9), 10, "Y"),
+        ("1-000000002", date(2024, 1, 9), 10, "Y"),
+    ]
