+++ projects/_01_ingest/unittest_data/ingest_test_file_schemas.py
@@ -377,3 +377,26 @@ class CleanCQCPIRSchema:
             StructField(CQCPIRClean.pir_submission_date_as_date, DateType(), True),
         ]
     )
+
+
+@dataclass
+class ValidatePIRRawData:
+    raw_cqc_pir_schema = StructType(
+        [
+            StructField(CQCPIR.location_id, StringType(), True),
+            StructField(Keys.import_date, StringType(), True),
+            StructField(CQCPIR.pir_people_directly_employed, StringType(), True),
+        ]
+    )
+
+
+@dataclass
+class ValidatePIRCleanedData:
+    cleaned_cqc_pir_schema = StructType(
+        [
+            StructField(CQCPIRClean.location_id, StringType(), True),
+            StructField(CQCPIRClean.cqc_pir_import_date, DateType(), True),
+            StructField(CQCPIRClean.pir_people_directly_employed, StringType(), True),
+            StructField(CQCPIRClean.care_home, StringType(), True),
+        ]
+    )
